Fixes a error when disguising as a horse with options

+++ src/me/libraryaddict/disguise/disguisetypes/watchers/HorseWatcher.java
@@ -88,7 +88,7 @@ public class HorseWatcher extends AgeableWatcher {
     }
 
     private void setFlag(int i, boolean flag) {
-        int j = (Integer) getValue(16, (byte) 0);
+        int j = (Byte) getValue(16, (byte) 0);
         if (flag) {
             setValue(16, j | i);
         } else {
